rTree implementation
r tree is implemented in BetterEdgeCrossingRTree.java
