+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "java.project.referencedLibraries": [
         "lib/**/*.jar",
-        "h:\\Downloads\\jar_files (1)\\rtree2-0.9.3.jar",
-        "h:\\Downloads\\jar_files (1)\\guava-mini-0.1.4.jar"
+        "h:\\Downloads\\jar_files (1)\\guava-mini-0.1.4.jar",
+        "h:\\Downloads\\jar_files (1)\\rtree2-0.9.3.jar"
     ]
 }
