+++ src/GraphXings/GraphXings.java
@@ -10,28 +10,28 @@ import GraphXings.Game.GameInstance.GameInstanceFactory;
 import GraphXings.Game.GameInstance.RandomCycleFactory;
 import GraphXings.Game.Match.NewMatch;
 import GraphXings.Game.Match.NewMatchResult;
-import GraphXings.NewFiles.NewMixingPlayer;
-import GraphXings.NewFiles.NewMixingPlayer.Strategy;
+import GraphXings.NewFiles.MixingPlayer;
+import GraphXings.NewFiles.MixingPlayer.Strategy;
 
 public class GraphXings {
     public static void main(String[] args) {
         // global matchup config
         int bestOutOf = 10;
         int sampleSize = 10;
-        double percentage = 0.995;
+        double percentage = 0.90;
         Strategy strategy = Strategy.Annealing;
-        NewPlayer player1 = new NewMixingPlayer("NewMixingPlayer", sampleSize, percentage, strategy);
-        NewPlayer player2 = new NewRandomPlayer("NewRandomPlayer");
+        NewPlayer player1 = new MixingPlayer("MixingPlayer", sampleSize, percentage, strategy);
+        NewPlayer player2 = new NewRandomPlayer("RandomPlayer");
 
         // type of game instanciation (random cycle -> true or constant -> false)
-        boolean cycleFactory = true;
+        boolean cycleFactory = false;
         // config to generate the graph from cycleFactory
-        int seed = 1010;
-        boolean includeMatchingEdges = true;
+        int seed = 108910;
+        boolean includeMatchingEdges = false;
         // config to generate the graph for the constant game only(!)
-        int numberOfVertices = 100;
-        int width = 100;
-        int height = 100;
+        int numberOfVertices = 1000;
+        int width = 1000;
+        int height = 1000;
         assert numberOfVertices <= (width * height) : "Graph not possible";
 
         GameInstanceFactory gif = null;
