+++ src/GraphXings/NewFiles/README.md
@@ -1,0 +1,1 @@
+Use NewMixingPlayer with this signature: public NewMixingPlayer(String name)
